MLK-19242 gpu: imx: dpu: framegen: Increase timeout when waiting for moving
It takes 33 milliseconds to scanout a frame when display refresh rate
is 30fps, and 42 milliseconds for 24fps.  So, if we wait for framegen
frame index moving, 30 milliseconds timeout value is not enough to
cover all reasonable display refresh rates.  50 milliseconds should not
be a bad choice.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
(cherry picked from commit 8564546882c940bb1852cbd132944b498f7ed102)