MLK-19365: crypto: tcrypt: Remove testing of hmac(crc32)
The kernel implementation of crc32 (crc32_generic.c)
accepts a key to set a seed. It is incompatible with the
kernel implementation of the crypto template hmac which
does not support keyed algorithms.

So it is not possible to load the algorithm hmac(crc32)
so remove it from tcrypt.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
(cherry picked from commit f283fdef31bb1af61079e421769a0c6de9bdb4f1)