MA-13207-3 [Android] the screen of first OS will be black once android insmod display module in xen
lvds0_en lvds1_en share the same gpio gpio1_6 on mek-8qm board.
the first os on xen will be black if add reset-gpios in android side.
the screen will be black if do not add reset-gpios in android side.
it can fix this issue when make it6263 as module.

Change-Id: I0180f5324fab411af32c9539798e2a1a659754cd
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>