MLK-21013 ARM64: dts: imx8qm: passthrough lvds pwm
Passthrough lvds pwm, otherwise dom0 will panic when reboot if
domu started and shutdown, because LVDS_1_PWM_0 is assigned
to domu and when domu shutdown, the resource will be powered
off, however dom0 still think it is powered on.

since lvds1_pwm is expected for domu, so let's passthrough it.

Signed-off-by: Peng Fan <peng.fan@nxp.com>