MA-14430 fix function definition compatibility build error with clang
the functions modified in this commit are assigned to the "set" member of
"struct kernel_param_ops" type variable. "set" points to a function with
the second parameter have const attribute, while the current definition
of these functions does not. This triggers build errors when compiled with
clang with default options.

Change-Id: Iab544358b67695b76b71db5b3fc6589086a68703
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>