MLK-22575-2: ASoC: fsl_easrc: disable PF_BYPASS_MODE
When run conversion in parrallel, one is down sampling, another
is up sampling, so one enables PF_BYPASS_MODE, another disable
PF_BYPASS_MODE, when the allocated slots in a same context
processor, there is noise in the result of down sampling.

If we alway disable PF_BYPASS_MODE, the issue can't be reproduced.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>